+++ api/src/controllers/client/index.ts
@@ -1,7 +1,6 @@
 // Copyright 2023 Touca, Inc. Subject to Apache-2.0 License.
 
-export { clientBatchNext } from './batchNext.js'
-export { clientElementList } from './elementList.js'
+export { clientBatchSeal } from './seal.js'
 export { clientOptions } from './options.js'
 export { clientSessionCreate } from './sessionCreate.js'
 export { clientSubmit } from './submit.js'