+++ api/src/controllers/client/seal.ts
@@ -9,9 +9,9 @@ import {
   ITeam,
   IUser
 } from '../../schemas/index.js'
-import { analytics, logger } from '../../utils/index.js'
+import { analytics, config, logger } from '../../utils/index.js'
 
-export async function ctrlBatchSeal(
+export async function clientBatchSeal(
   req: Request,
   res: Response,
   next: NextFunction
@@ -22,12 +22,12 @@ export async function ctrlBatchSeal(
   const batch = res.locals.batch as IBatchDocument
   const tuple = [team.slug, suite.slug, batch.slug].join('/')
   logger.debug('%s: %s: sealing', user.username, tuple)
-  if (batch.sealedAt) {
-    logger.info('%s: %s: already sealed', user.username, tuple)
-    return res.status(204).send()
+  if (!batch.sealedAt) {
+    await batchSeal(team, suite, batch)
+    analytics.add_activity('batch:sealed', user)
   }
-  await batchSeal(team, suite, batch)
-  analytics.add_activity('batch:sealed', user)
   logger.info('%s: %s: sealed', user.username, tuple)
-  return res.status(204).send()
+  return res.status(200).json({
+    link: `${config.webapp.root}/~/${team.slug}/${suite.slug}/${batch.slug}`
+  })
 }