+++ sdk/cpp/src/client.cpp
@@ -169,13 +169,13 @@ bool ClientImpl::post() const {
   return ret;
 }
 
-void ClientImpl::seal() const {
+std::string ClientImpl::seal() const {
   if (!_configured || _options.offline) {
     throw touca::detail::runtime_error(
         "client is not configured to contact the server");
   }
   const auto& response = _transport->post(
-      touca::detail::format("/batch/{}/{}/{}/seal2", _options.team,
+      touca::detail::format("/client/seal/{}/{}/{}", _options.team,
                             _options.suite, _options.version));
   if (response.status == -1) {
     throw touca::detail::runtime_error(
@@ -184,10 +184,21 @@ void ClientImpl::seal() const {
   if (response.status == 403) {
     throw touca::detail::runtime_error("client is not authenticated");
   }
-  if (response.status != 204) {
+  if (response.status != 200) {
     throw touca::detail::runtime_error(
         touca::detail::format("failed to seal version: {}", response.status));
   }
+  rapidjson::Document parsed;
+  if (parsed.Parse<0>(response.body.c_str()).HasParseError()) {
+    throw touca::detail::runtime_error("failed to parse server response");
+  }
+  if (parsed.IsObject()) {
+    const auto& members = parsed.GetObject();
+    if (members.HasMember("link") && members["link"].IsString()) {
+      return members["link"].GetString();
+    }
+  }
+  throw touca::detail::runtime_error("server response was unexpected");
 }
 
 bool ClientImpl::has_last_testcase() const {