+++ sdk/cpp/src/runner.cpp
@@ -205,7 +205,9 @@ void Printer::print_progress(const unsigned index, const Status status,
 }
 
 void Printer::print_footer(const Statistics& stats, Timer& timer,
-                           const unsigned suiteSize) {
+                           const Workflow workflow,
+                           const RunnerOptions& options,
+                           const std::string& web_link) {
   const auto duration = timer.count("__workflow__") / 1000.0;
   const auto report = [&](const Status state, const fmt::terminal_color color,
                           const std::string& name) {
@@ -218,8 +220,17 @@ void Printer::print_footer(const Statistics& stats, Timer& timer,
   report(Status::Pass, fmt::terminal_color::green, "passed");
   report(Status::Skip, fmt::terminal_color::yellow, "skipped");
   report(Status::Fail, fmt::terminal_color::red, "failed");
-  print("{} total\n", suiteSize);
+  print("{} total\n", static_cast<unsigned int>(workflow.testcases.size()));
   print("Time:       {:.2f} s\n", duration);
+  if (!web_link.empty()) {
+    print("Link:       {}\n", web_link);
+  }
+  if (options.save_binary || options.save_json) {
+    const auto& results_dir =
+        touca::filesystem::path(options.output_directory) / workflow.suite /
+        workflow.version;
+    print("Results:    {}\n", results_dir.string());
+  }
 }
 
 void Runner::run_workflows() {
@@ -284,12 +295,9 @@ void Runner::run_workflow(const Workflow& workflow) {
     run_testcase(workflow, testcase, index++);
   }
   timer.toc("__workflow__");
-  printer.print_footer(stats, timer,
-                       static_cast<unsigned int>(workflow.testcases.size()));
-
-  if (!options.offline && stats.count(Status::Pass) != 0) {
-    touca::seal();
-  }
+  std::string web_link =
+      options.offline || stats.count(Status::Pass) == 0 ? "" : touca::seal();
+  printer.print_footer(stats, timer, workflow, options, web_link);
 }
 
 void Runner::run_testcase(const Workflow& workflow, const std::string& testcase,