+++ sdk/python/touca/_runner.py
@@ -185,9 +185,8 @@ def _run_workflow(options: dict):
         Client.instance().forget_testcase(testcase)
 
     timer.toc("__workflow__")
-    printer.print_footer(stats, timer, options)
-    if not options.get("offline"):
-        Client.instance().seal()
+    web_link = "" if options.get("offline") else Client.instance().seal()
+    printer.print_footer(stats, timer, options, web_link)
 
 
 def run_workflows(opts):