+++ sdk/python/Changelog.md
@@ -2,6 +2,14 @@
 
 ## v1.8.6
 
+Breaking Changes:
+
+- This SDK only works with Touca server v1.9 and later
+
+Features:
+
+- Test runner output should include link to results on the server
+
 ## v1.8.5
 
 Features: